~ clean up type hierarchy and codecs
