+++ opc-ua-sdk/integration-tests/src/test/java/org/eclipse/milo/opcua/sdk/client/DynamicMatrixTestTypeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 the Eclipse Milo Authors
+ * Copyright (c) 2025 the Eclipse Milo Authors
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -14,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.eclipse.milo.opcua.sdk.client.nodes.UaVariableNode;
-import org.eclipse.milo.opcua.sdk.core.types.DynamicStruct;
+import org.eclipse.milo.opcua.sdk.core.types.DynamicStructType;
 import org.eclipse.milo.opcua.sdk.test.AbstractClientServerTest;
 import org.eclipse.milo.opcua.sdk.test.MatrixTestType;
 import org.eclipse.milo.opcua.stack.core.UaException;
@@ -38,7 +38,7 @@ public class DynamicMatrixTestTypeTest extends AbstractClientServerTest {
     ExtensionObject xo = (ExtensionObject) value.getValue().getValue();
     assert xo != null;
 
-    DynamicStruct decoded = (DynamicStruct) xo.decode(client.getDynamicEncodingContext());
+    DynamicStructType decoded = (DynamicStructType) xo.decode(client.getDynamicEncodingContext());
     assertEquals(
         MatrixTestType.TYPE_ID,
         decoded.getTypeId().absolute(client.getNamespaceTable()).orElseThrow());
