+++ opc-ua-sdk/sdk-core/src/main/java/org/eclipse/milo/opcua/sdk/core/types/DynamicOptionSetType.java
@@ -18,23 +18,43 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.StringJoiner;
 import org.eclipse.milo.opcua.sdk.core.typetree.DataType;
+import org.eclipse.milo.opcua.stack.core.types.UaStructuredType;
 import org.eclipse.milo.opcua.stack.core.types.builtin.ByteString;
+import org.eclipse.milo.opcua.stack.core.types.builtin.ExpandedNodeId;
 import org.eclipse.milo.opcua.stack.core.types.builtin.LocalizedText;
 import org.eclipse.milo.opcua.stack.core.types.structured.EnumDefinition;
 import org.eclipse.milo.opcua.stack.core.types.structured.EnumField;
 import org.eclipse.milo.opcua.stack.core.util.Lazy;
 import org.jspecify.annotations.Nullable;
 
-public class DynamicOptionSet extends DynamicStruct {
+public final class DynamicOptionSetType extends DynamicType implements UaStructuredType {
 
   private final Lazy<Map<Integer, EnumField>> lazyFieldMap = new Lazy<>();
 
-  public DynamicOptionSet(DataType dataType) {
+  private final DataType dataType;
+  private final LinkedHashMap<String, Object> members;
+
+  public DynamicOptionSetType(DataType dataType) {
     this(dataType, new LinkedHashMap<>());
   }
 
-  public DynamicOptionSet(DataType dataType, LinkedHashMap<String, Object> members) {
-    super(dataType, members);
+  public DynamicOptionSetType(DataType dataType, LinkedHashMap<String, Object> members) {
+    this.dataType = dataType;
+    this.members = members;
+  }
+
+  @Override
+  public ExpandedNodeId getTypeId() {
+    return dataType.getNodeId().expanded();
+  }
+
+  @Override
+  public DataType getDataType() {
+    return dataType;
+  }
+
+  public LinkedHashMap<String, Object> getMembers() {
+    return members;
   }
 
   public ByteString getValue() {
@@ -76,7 +96,7 @@ public class DynamicOptionSet extends DynamicStruct {
         () -> {
           Map<Integer, EnumField> fieldMap = Collections.synchronizedMap(new HashMap<>());
 
-          EnumDefinition definition = (EnumDefinition) getDataType().getDataTypeDefinition();
+          EnumDefinition definition = (EnumDefinition) dataType.getDataTypeDefinition();
           assert definition != null;
 
           EnumField[] fields = requireNonNullElse(definition.getFields(), new EnumField[0]);
@@ -91,7 +111,7 @@ public class DynamicOptionSet extends DynamicStruct {
 
   @Override
   public String toString() {
-    return new StringJoiner(", ", DynamicOptionSet.class.getSimpleName() + "[", "]")
+    return new StringJoiner(", ", DynamicOptionSetType.class.getSimpleName() + "[", "]")
         .add("value=" + toBitString(getValue()))
         .add("validBits=" + toBitString(getValidBits()))
         .toString();
