+++ opc-ua-sdk/sdk-core/src/main/java/org/eclipse/milo/opcua/sdk/core/types/codec/DynamicOptionSetCodec.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 the Eclipse Milo Authors
+ * Copyright (c) 2025 the Eclipse Milo Authors
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -8,8 +8,9 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.milo.opcua.sdk.core.types;
+package org.eclipse.milo.opcua.sdk.core.types.codec;
 
+import org.eclipse.milo.opcua.sdk.core.types.DynamicOptionSetType;
 import org.eclipse.milo.opcua.sdk.core.typetree.DataType;
 import org.eclipse.milo.opcua.stack.core.UaSerializationException;
 import org.eclipse.milo.opcua.stack.core.encoding.EncodingContext;
@@ -18,7 +19,7 @@ import org.eclipse.milo.opcua.stack.core.encoding.UaDecoder;
 import org.eclipse.milo.opcua.stack.core.encoding.UaEncoder;
 import org.eclipse.milo.opcua.stack.core.types.builtin.ByteString;
 
-class DynamicOptionSetCodec extends GenericDataTypeCodec<DynamicOptionSet> {
+class DynamicOptionSetCodec extends GenericDataTypeCodec<DynamicOptionSetType> {
 
   private final DataType dataType;
 
@@ -27,17 +28,17 @@ class DynamicOptionSetCodec extends GenericDataTypeCodec<DynamicOptionSet> {
   }
 
   @Override
-  public Class<DynamicOptionSet> getType() {
-    return DynamicOptionSet.class;
+  public Class<DynamicOptionSetType> getType() {
+    return DynamicOptionSetType.class;
   }
 
   @Override
-  public DynamicOptionSet decodeType(EncodingContext context, UaDecoder decoder)
+  public DynamicOptionSetType decodeType(EncodingContext context, UaDecoder decoder)
       throws UaSerializationException {
     ByteString value = decoder.decodeByteString("Value");
     ByteString validBits = decoder.decodeByteString("ValidBits");
 
-    var optionSet = new DynamicOptionSet(dataType);
+    var optionSet = new DynamicOptionSetType(dataType);
     optionSet.setValue(value);
     optionSet.setValidBits(validBits);
 
@@ -45,7 +46,7 @@ class DynamicOptionSetCodec extends GenericDataTypeCodec<DynamicOptionSet> {
   }
 
   @Override
-  public void encodeType(EncodingContext context, UaEncoder encoder, DynamicOptionSet value)
+  public void encodeType(EncodingContext context, UaEncoder encoder, DynamicOptionSetType value)
       throws UaSerializationException {
     encoder.encodeByteString("Valid", value.getValue());
     encoder.encodeByteString("ValidBits", value.getValidBits());
